feat: Restrict aggregation to start and end dates
When performing an aggregation on an indicator, we first get the most
recent raw measure date used to perform the last aggregation: this
provides the starting date of the aggregation.
Then, we select all the tuples of this indicator, starting from this date,
sorted by the `last_updated` in desc. This first tuple provides the
ending date.
Finally, we query the raw measures included in this time interval.