feat: Support partial aggregates
When a new aggregation is performed on raw measures, an entry might already
exists in the Aggregation table.
If this is the case, we compute a new aggregation between the existing
tuple and the new one.
Note this is possible because we only support aggregate functions that
can be run on partitioned data (distributive and algebric). This would
not be possible with holistic functions such as median.