feat: Add flask-migrate
This is very useful to deal with database migrations: after a change in
the database structure, one can generate migration script with `flask db
migrate`. Then, this script can be used to automatically migrate thanks
to `flask db upgrade`.
Note it is also possible to revert a migration with `flask db downgrade`