feat: Use test database for pytest
During development, using the same database for unit tests can be
problematic because we erase it at each run.
Thus, we introduce the possibility to use a distinct test database in the
config.yml.
We pass a TESTING env var from pytest, as defined in the SQLAlchemy
config: https://flask.palletsprojects.com/en/2.0.x/config/#TESTING