feat: Add the possibility to aggregate quartiles
We used to restrict aggregates to functions that can be partitioned,
i.e. that can be computed based on previous aggregate, without requiring
raw values.
We introduce here the possibility to compute quartiles, which breaks
this assumption: when an existing aggregate must be updated, all the
previous measures are queried to compute the quartiles from scratch.

By default, a measure does not compute quartile. It must be expressed in
its definition by setting `with_quartile: true`.

Note this will complicate a future purge mechanism: while we could
safely remove raw measures once aggregated, this is no longer possible
with quartiles.

feat: Migrate database