feat: Add CLI to delete existing aggregations
This removes aggregates for a specified measure when their start_date is
superior to the given date.
It also changes the last_aggregated_measure_date so aggregations can
be computed computed again from this date.
Note this can have side-effects: suppose a measure with a start_date:
01/10/21 is inserted in december, and we wish to remove all aggregations
for this month.
The next aggregation run will include the 01/10/21 measure, while the
aggregate for october still exists, making this measure counted twice.

Therefore, this feature should be used with care.